Normalized the library within namespaces, and added Composer autoload.
