+++ protocol/SmppException.php
@@ -1,0 +1,12 @@
+<?php
+namespace Phpsmpp\Protocol;
+/**
+ * Created by PhpStorm.
+ * User: nfargere
+ * Date: 14/10/2016
+ * Time: 09:45
+ */
+class SmppException extends \RuntimeException
+{
+
+}
