+++ transport/TException.php
@@ -1,0 +1,3 @@
+<?php
+namespace Phpsmpp\Transport;
+class TException extends \Exception { };
