--- transport/texception.class.php
@@ -1,2 +1,0 @@
-<?php
-class TException extends Exception { };
